:art: Fix some type names
